+++ packages/wrapper/web/src/themeStrategy/themes/editor-common.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-len */
 /* eslint-disable @typescript-eslint/camelcase */
 import { adaptForeground, hexToRgbA } from './utils';
 
@@ -5,159 +6,184 @@ export default function editorCommon(colors: PaletteColors) {
   const actionColor = adaptForeground(colors.actionColor);
   const blockActionColorSettings = {
     cursor: 'default',
-    boxShadow: `0 0 0 3px ${colors.actionColor} !important`,
+    boxShadow: `0 0 0 3px ${colors.actionColor}`,
   };
-  const sliderTrack = { background: `${actionColor} !important` };
+  const toolbarButtonStyle = {
+    color: actionColor,
+    fill: actionColor,
+  };
+  const sliderTrack = { background: actionColor };
   const thumb = { ...sliderTrack, border: `4px solid ${actionColor}` };
+  //button.scss
+  const buttonsFooterStyle = {
+    '& $button_primary': {
+      backgroundColor: actionColor,
+    },
+    '& $button_primary:hover:not([disabled])': {
+      backgroundColor: hexToRgbA(actionColor, 0.8),
+    },
+    '& $button_primary:disabled': {
+      backgroundColor: 'rgba(0, 0, 0, 0.4)',
+    },
+    '& $button_secondary': {
+      color: actionColor,
+      borderColor: actionColor,
+    },
+    '& $button_secondary:hover': {
+      color: hexToRgbA(actionColor, 0.6),
+      borderColor: hexToRgbA(actionColor, 0.6),
+    },
+    '& $button_secondary:disabled': {
+      backgroundColor: 'rgba(0, 0, 0, 0.4)',
+    },
+  };
   return {
     //block focus
     hasFocus: blockActionColorSettings,
-    pluginContainer: { '&:hover': blockActionColorSettings },
+    pluginContainer: {},
+    pluginContainerWrapper: {
+      '&$pluginContainer:hover': blockActionColorSettings,
+      '&$pluginContainer&$hasFocus': blockActionColorSettings,
+    },
 
     //selection-list.scss
-    selectionListOption: {
-      '&:hover': {
-        backgroundColor: `${hexToRgbA(actionColor, 0.05)} !important`,
+    selectionList: {
+      '& $selectionListOption:hover': {
+        backgroundColor: hexToRgbA(actionColor, 0.05),
       },
-    },
-    selectionListOption_selected: {
-      color: actionColor,
-      backgroundColor: `${hexToRgbA(actionColor, 0.1)} !important`,
-      '&$selectionListOption:hover': {
+      '& $selectionListOption$selectionListOption_selected': {
+        color: actionColor,
+        backgroundColor: hexToRgbA(actionColor, 0.1),
+      },
+      '& $selectionListOption_selected$selectionListOption:hover': {
         backgroundColor: hexToRgbA(actionColor, 0.1),
       },
     },
+    selectionListOption: {},
+    selectionListOption_selected: {},
 
     //checkbox.scss
-    checkbox_icon: {
-      border: `solid 1px ${actionColor} !important`,
-      '&:hover': {
-        color: `${hexToRgbA(actionColor, 0.1)} !important`,
+    checkbox_wrapper: {
+      '& $checkbox_icon': {
+        border: `solid 1px ${actionColor}`,
+      },
+      '& $checkbox_icon:hover': {
+        color: hexToRgbA(actionColor, 0.1),
+      },
+      '& $checkbox_infoIcon:hover': {
+        color: actionColor,
+      },
+      '& $checkbox_inputLabel:hover $checkbox_icon_unchecked': {
+        backgroundColor: hexToRgbA(actionColor, 0.1),
+      },
+      '& $checkbox_icon_checked': {
+        backgroundColor: actionColor,
       },
     },
-    checkbox_infoIcon: {
-      '&:hover': {
-        color: `${actionColor} !important`,
-      },
-    },
-    checkbox_icon_checked: {
-      backgroundColor: `${actionColor} !important`,
-    },
-    checkbox_inputLabel: {
-      '&:hover $checkbox_icon_unchecked': {
-        backgroundColor: `${hexToRgbA(actionColor, 0.1)} !important`,
-      },
-    },
+    checkbox_icon: {},
+    checkbox_infoIcon: {},
+    checkbox_icon_checked: {},
+    checkbox_inputLabel: {},
     checkbox_icon_unchecked: {},
 
     //tabs.scss
-    tabs_panel: {
-      padding: '24px 24px 30px !important',
+    tabs_panel: {},
+    tabs: {
+      '& $tabs_panel': {
+        padding: '24px 24px 30px',
+      },
     },
-    tabs_headers_option_selected: {
-      borderBottom: `solid 3px ${actionColor} !important`,
+    tabs_headers_option_selected: {},
+    tabs_headers: {
+      '& $tabs_headers_option_selected': {
+        borderBottom: `solid 3px ${actionColor}`,
+      },
     },
 
     //button.scss
-    button_primary: {
-      backgroundColor: `${actionColor} !important`,
-      '&:hover:not([disabled])': {
-        backgroundColor: `${hexToRgbA(actionColor, 0.8)} !important`,
-      },
-      '&:disabled': {
-        backgroundColor: 'rgba(0, 0, 0, 0.4) !important',
-      },
-    },
-    button_secondary: {
-      color: actionColor,
-      borderColor: `${actionColor} !important`,
-      '&:hover': {
-        color: hexToRgbA(actionColor, 0.6),
-        borderColor: `${hexToRgbA(actionColor, 0.6)} !important`,
-      },
-      '&:disabled': {
-        backgroundColor: 'rgba(0, 0, 0, 0.4) !important',
-      },
-    },
+    button_primary: {},
+    button_secondary: {},
+    settingsPanel_footer_fixed: buttonsFooterStyle, //gallery, map...
+    button_inputModal_modal_footer: buttonsFooterStyle, //button plugin
+    video_modal_container_big: buttonsFooterStyle, //video plugin
+    video_modal_container_small: buttonsFooterStyle, //video plugin
 
     //slider.scss
-    slider: {
-      '&::-webkit-slider-runnable-track': sliderTrack,
-      '&::-webkit-slider-thumb': thumb,
+    sliderWithInput_content: {
+      '& $slider::-webkit-slider-runnable-track': sliderTrack,
+      '& $slider::-webkit-slider-thumb': thumb,
     },
+    slider: {},
 
     //radio-group.scss
     radioGroup_button: {},
-    radioGroup_input: {
-      '&:checked + $radioGroup_button': {
-        '&::after': {
-          backgroundColor: `${actionColor} !important`,
-        },
+    radioGroup_input: {},
+    radioGroup: {
+      '& $radioGroup_input:checked + $radioGroup_button::after': {
+        backgroundColor: actionColor,
       },
     },
 
     //color-picker.scss
-    colorPicker_add_color_label: {
-      color: `${actionColor} !important`,
+    colorPicker_add_color_button: {
+      '& $colorPicker_add_color_label': {
+        color: actionColor,
+      },
     },
+    colorPicker_add_color_label: {},
+    colorPicker_reset_color_button: {
+      '& $colorPicker_reset_color_label': {
+        color: actionColor,
+      },
+    },
+    colorPicker_reset_color_label: {},
 
     //side-toolbar.scss + toolbar-button.scss
     //editor + editor-common
-    sideToolbar: {
-      '& button:hover:not([disabled]) svg, button:focus:not([disabled]) svg': {
-        color: `${actionColor} !important`,
-        fill: `${actionColor} !important`,
-      },
-      '& label:hover:not([disabled]) svg, label:focus:not([disabled]) svg': {
-        color: `${actionColor} !important`,
-        fill: `${actionColor} !important`,
-      },
+    buttonsWrapper: {
+      '& $buttonWrapper:hover button:not([disabled]) svg': toolbarButtonStyle,
+      '& $buttonWrapper:focus button:not([disabled]) svg': toolbarButtonStyle,
+      '& $buttonWrapper:hover label:not([disabled]) svg': toolbarButtonStyle,
+      '& $buttonWrapper:focus label:not([disabled]) svg': toolbarButtonStyle,
+    },
+    buttonWrapper: {
+      //plugin-toolbar-button.scss
+      '& $pluginToolbarButton:not($pluginToolbarButton_disabled)$pluginToolbarButton_active': toolbarButtonStyle,
+      '& $pluginToolbarButton:not($pluginToolbarButton_disabled):focus svg': toolbarButtonStyle,
+      '& $pluginToolbarButton:not($pluginToolbarButton_disabled):hover': toolbarButtonStyle,
     },
 
     //inline-toolbar-button.scss
-    inlineToolbarButton_active: {},
-    inlineToolbarButton_icon: {
-      '&:hover': {
-        color: `${actionColor} !important`,
-        '& svg': {
-          fill: `${actionColor} !important`,
-        },
+    inlineToolbarButton_wrapper: {
+      '& $inlineToolbarButton_icon:hover': {
+        color: actionColor,
+      },
+      '& $inlineToolbarButton_icon:hover svg': {
+        fill: actionColor,
       },
     },
+    inlineToolbarButton_icon: {},
 
     //plugin-toolbar-button.scss
     pluginToolbarButton_disabled: {},
-    pluginToolbarButton_active: {
-      color: `${actionColor} !important`,
-      fill: `${actionColor} !important`,
-    },
+    pluginToolbarButton_active: {},
     pluginToolbarButton: {
       '&:not($pluginToolbarButton_disabled):hover >:not($Dropdown-root) svg': {
-        color: `${actionColor} !important`,
-        fill: `${actionColor} !important`,
+        fill: actionColor,
       },
-      '&:not($pluginToolbarButton_disabled):focus svg': {
-        color: `${actionColor} !important`,
-        fill: `${actionColor} !important`,
-      },
-      '&:not($pluginToolbarButton_disabled)$pluginToolbarButton_active svg': {
-        color: `${actionColor} !important`,
-        fill: `${actionColor} !important`,
-      },
-      '&:not($pluginToolbarButton_disabled):hover': {},
     },
 
     //dropdown.scss
-    'Dropdown-control': {
-      '&:hover': {
-        color: `${actionColor} !important`,
+    'Dropdown-control': {},
+    'Dropdown-option': {},
+    'Dropdown-root': {
+      '& $Dropdown-control:hover': {
+        color: actionColor,
       },
-    },
-    'Dropdown-option': {
-      '&:hover': {
-        color: `${actionColor} !important`,
+      '& $Dropdown-option:hover': {
+        color: actionColor,
       },
+      '& $Dropdown-option:hover svg': toolbarButtonStyle,
     },
-    'Dropdown-root': {},
   };
 }