+++ rollup.externals.js
@@ -8,6 +8,7 @@ const externals = [
   'wix-rich-content-editor-common',
   'wix-rich-content-common',
   'react-i18next',
+  /^punycode$/,
   /^@wix\/draft-js$/,
   /^jss$/, //issue with ESM in CJS
   /^jss-preset-default$/, //issue with ESM in CJS